[mlir] Add option to run CSE between greedy rewriter iterations
The greedy pattern rewrite driver previously only deduplicated constant ops
between iterations (via the operation folder). Structurally identical
non-constant subexpressions remained distinct SSA values, blocking fold
patterns that only fire when operands match. Reaching the true fixpoint
required chaining an external `cse,canonicalize,...` pipeline.

Add an opt-in `cseBetweenIterations` flag on `GreedyRewriteConfig` that runs
full CSE on the scoped region after each pattern-application iteration, and
surface it as a `cse-between-iterations` option on the canonicalizer pass.
Off by default to preserve existing performance characteristics.

To let the greedy driver (in MLIRTransformUtils) call into CSE without
creating a layering cycle with MLIRTransforms, the CSE driver implementation
moves to `Utils/CSE.cpp`; the CSE pass in `Transforms/CSE.cpp` becomes a thin
wrapper over the public API. A region-scoped overload of
`eliminateCommonSubExpressions` is added for use by the driver.
