+++ mlir/include/mlir/Transforms/CSE.h
@@ -13,18 +13,34 @@
 #ifndef MLIR_TRANSFORMS_CSE_H_
 #define MLIR_TRANSFORMS_CSE_H_
 
+#include <cstdint>
+
 namespace mlir {
 
 class DominanceInfo;
 class Operation;
+class Region;
 class RewriterBase;
 
 /// Eliminate common subexpressions within the given operation. This transform
 /// looks for and deduplicates equivalent operations.
 ///
-/// `changed` indicates whether the IR was modified or not.
+/// `changed` indicates whether the IR was modified or not. `numCSE` and
+/// `numDCE` receive counts of operations deduplicated and dead operations
+/// erased, respectively.
 void eliminateCommonSubExpressions(RewriterBase &rewriter,
                                    DominanceInfo &domInfo, Operation *op,
+                                   bool *changed = nullptr,
+                                   int64_t *numCSE = nullptr,
+                                   int64_t *numDCE = nullptr);
+
+/// Eliminate common subexpressions within the given region.
+///
+/// `changed` indicates whether the IR was modified or not. Statistics are not
+/// reported through this overload; use the `Operation *` overload when CSE /
+/// DCE counts are needed.
+void eliminateCommonSubExpressions(RewriterBase &rewriter,
+                                   DominanceInfo &domInfo, Region &region,
                                    bool *changed = nullptr);
 
 } // namespace mlir
